feature:Creacion de evento Rabbit

+++ ordersrvc/src/main/java/py/ecommerce/order/dto/OrderCreateEvent.java
@@ -1,6 +1,7 @@
 package py.ecommerce.order.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
-public record OrderCreateEvent(Long orderId, String customerEmail, Double totalAmount) implements Serializable {
+public record OrderCreateEvent(Long orderId, String customerEmail, BigDecimal totalAmount) implements Serializable {
 }
